requirements.txt
pyyaml
pandas
kagglehub